DOCS update deprecation notice with more helpful pointers
People reading the deprecation notice were sometimes confused about what
steps to take next - as this module was replaced with a recipe after
it's components were split out, I think it fair to relay this
information to the user at the same time as the notice, along with hints
in the composer manifest. Hopefully this gives better direction on what
to do next in order to install an up to date set of packages for
elemental on newer SilverStripe versions.